requirements.txt
pytest==7.1.1
appium-python-client==2.1.4
selenium==4.1.3